Simplify voter storage; remove key storage.
