+++ lib/blocks.js
@@ -5,7 +5,6 @@
 
 const _ = require('lodash');
 const _cache = require('./cache');
-const _storage = require('./storage');
 const _util = require('./util');
 const bedrock = require('bedrock');
 const {callbackify, BedrockError} = bedrock.util;
