+++ lib/config.js
@@ -13,6 +13,7 @@ const os = require('os');
 const cfg = config['ledger-consensus-continuity'] = {};
 
 cfg.routes = {};
+// FIXME: remove `root`, unused
 cfg.routes.root = '/consensus/continuity2017/voters/:voterId';
 cfg.routes.events = cfg.routes.root + '/events';
 cfg.routes.eventsQuery = cfg.routes.root + '/events-query';
