+++ lib/events.js
@@ -13,7 +13,6 @@ const bedrock = require('bedrock');
 const brLedgerNode = require('bedrock-ledger-node');
 const {callbackify, BedrockError} = bedrock.util;
 const {config} = bedrock;
-const continuityStorage = require('./storage');
 const database = require('bedrock-mongodb');
 const jsonld = bedrock.jsonld;
 const logger = require('./logger');
