+++ lib/index.js
@@ -31,7 +31,6 @@ api._gossip = require('./gossip');
 api._hasher = brLedgerNode.consensus._hasher;
 api._server = require('./server');
 api._voters = require('./voters');
-api._storage = require('./storage');
 api._worker = require('./worker');
 
 // expose external APIs
