+++ lib/server.js
@@ -24,8 +24,6 @@ require('bedrock-permission');
 
 require('./config');
 
-const storage = require('./storage');
-
 // module API
 const api = {};
 module.exports = api;
@@ -43,7 +41,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
     get: callbackify(async (req, res) => {
       const voterId = config.server.baseUri +
         '/consensus/continuity2017/voters/' + req.params.voterId;
-      const voter = await storage.voters.get({voterId, publicKey: true});
+      const voter = await _voters.storage.get({voterId, publicKey: true});
       return {
         // bedrock-ledger-continuity
         '@context': config.constants.WEB_LEDGER_CONTEXT_V1_URL,
@@ -78,8 +76,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
         '/consensus/continuity2017/voters/' + req.params.voterId;
       const eventHash = req.query.id;
 
-      const voter = await storage.voters.get({voterId});
-      //const ledgerNodeId = await _voters.getLedgerNodeId(voterId);
+      const voter = await _voters.storage.get({voterId});
       const ledgerNode = await brLedgerNode.get(null, voter.ledgerNodeId);
       const {consensusMethod} = ledgerNode.consensus;
       if(consensusMethod !== 'Continuity2017') {
