--- lib/storage.js
@@ -1,12 +1,0 @@
-/*!
- * Storage for Continuity2017 consensus method.
- *
- * Copyright (c) 2017-2018 Digital Bazaar, Inc. All rights reserved.
- */
-'use strict';
-
-// module API
-const api = {};
-module.exports = api;
-
-api.voters = require('./storage-voters');
