+++ lib/voters.js
@@ -10,14 +10,15 @@ const cache = require('bedrock-redis');
 const {callbackify, BedrockError} = bedrock.util;
 const {config} = bedrock;
 const chloride = require('chloride');
+const database = require('bedrock-mongodb');
 const jsigs = require('jsonld-signatures')();
+const logger = require('./logger');
 const multicodec = require('multicodec');
 const multibase = require('multibase');
+const {promisify} = require('util');
 const URL = require('url');
 
 jsigs.use('jsonld', bedrock.jsonld);
-
-const _storage = require('./storage-voters');
 
 require('./config');
 
@@ -27,6 +28,19 @@ const KEY_GENERATION_TIME = 50;
 // module API
 const api = {};
 module.exports = api;
+
+bedrock.events.on('bedrock-mongodb.ready', async () => {
+  await promisify(database.openCollections)(['continuity2017_voter']);
+  await promisify(database.createIndexes)([{
+    collection: 'continuity2017_voter',
+    fields: {'voter.id': 1},
+    options: {unique: true, background: false}
+  }, {
+    collection: 'continuity2017_voter',
+    fields: {'voter.ledgerNodeId': 1},
+    options: {unique: true, background: false}
+  }]);
+});
 
 /**
  * Get the voter used for the given ledger node ID. If no such voter exists,
@@ -61,7 +75,7 @@ api.get = callbackify(async (
   let tries = 0;
   while(tries < 100) {
     try {
-      voter = await _storage.get({ledgerNodeId, privateKey, publicKey});
+      voter = await api.storage.get({ledgerNodeId, privateKey, publicKey});
       break;
     } catch(e) {
       if(e.name !== 'NotFoundError') {
@@ -120,7 +134,7 @@ api.getLedgerNodeId = callbackify(async (voterId) => {
     return result;
   }
 
-  const {ledgerNodeId} = await _storage.get({voterId});
+  const {ledgerNodeId} = await api.storage.get({voterId});
 
   // update cache
   await cache.client.set(ledgerNodeKey, ledgerNodeId);
@@ -192,6 +206,116 @@ api.getPeerIds = callbackify(async ({creatorId, ledgerNode, limit}) => {
   return collection.distinct('meta.continuity2017.creator', query);
 });
 
+api.storage = {};
+
+/**
+ * Adds a new voter. Each individual ledger node can be a "voter" and
+ * needs voter information including an identifier, a cryptographic key,
+ * and a ledger agent. The ledger agent URL is not stored in the database,
+ * but computed from the configuration using the ledger agent ID.
+ *
+ * FIXME: move private key storage to HSM/SSM.
+ *
+ * @param voter the ledger node's voter information:
+ *          id the ID of the voter.
+ *          ledgerNodeId the ID of the ledger node.
+ *          publicKey the voter's public key.
+ *          privateKey the voter's private key.
+ *
+ * @return a Promise that resolves to the voter record.
+ */
+api.storage.add = callbackify(async (voter) => {
+  logger.verbose('adding voter', voter.id);
+
+  // create the record
+  const now = Date.now();
+  const record = {
+    meta: {
+      created: now,
+      updated: now
+    },
+    voter
+  };
+  const collection = database.collections.continuity2017_voter;
+
+  try {
+    return (await collection.insert(record, database.writeOptions)).ops[0];
+  } catch(e) {
+    if(database.isDuplicateError(e)) {
+      // TODO: how to handle key rotation?
+      throw new BedrockError(
+        'The voter information for the given ledger node already exists.',
+        'DuplicateError', {ledgerNodeId: voter.ledgerNodeId}, e);
+    }
+    throw e;
+  }
+
+  return record;
+});
+
+/**
+ * Gets the voter information for the given ledger node ID or voter ID.
+ *
+ * @param options the options to use:
+ *          [ledgerNodeId] the ID of the ledger node.
+ *          [voterId] the ID of the voter.
+ *
+ * @return a Promise that resolves to the voter.
+ */
+api.storage.get = callbackify(async (
+  {ledgerNodeId, privateKey = false, publicKey = false, voterId}) => {
+  if(!(ledgerNodeId || voterId)) {
+    throw new Error('"ledgerNodeId" or "voterId" must be set.');
+  }
+
+  const query = {'meta.deleted': {$exists: false}};
+  if(ledgerNodeId) {
+    query['voter.ledgerNodeId'] = ledgerNodeId;
+  } else {
+    query['voter.id'] = voterId;
+  }
+
+  const collection = database.collections.continuity2017_voter;
+  const projection = {
+    _id: 0,
+    'voter.id': 1,
+    'voter.ledgerNodeId': 1,
+  };
+  if(publicKey) {
+    // FIXME: get `publicKeyBase58` from voter ID; does not need to
+    // be stored in database
+    Object.assign(projection, {
+      'voter.publicKey.id': 1,
+      'voter.publicKey.publicKeyBase58': 1
+    });
+  }
+  if(privateKey) {
+    // the private key and the associated `id` resides in `publicKey`
+    Object.assign(projection, {
+      'voter.publicKey.id': 1,
+      'voter.publicKey.privateKey.privateKeyBase58': 1
+    });
+  }
+  const record = await collection.findOne(query, projection);
+  if(!record) {
+    const details = {
+      httpStatusCode: 400,
+      public: true
+    };
+    if(ledgerNodeId) {
+      details.ledgerNodeId = ledgerNodeId;
+    } else {
+      details.voterId = voterId;
+    }
+
+    throw new BedrockError(
+      'Voter information not found for the given ledger node.',
+      'NotFoundError', details);
+  }
+
+  return record.voter;
+});
+
 async function _createVoter(ledgerNodeId) {
   const kp = await _generateEd25519Key();
 
@@ -213,7 +337,7 @@ async function _createVoter(ledgerNodeId) {
       }
     }
   };
-  const result = await _storage.add(voter);
+  const result = await api.storage.add(voter);
   return result.voter;
 }
